bugreport: move to tools
