+++ .pipelines/ci/templates/build-powertoys-steps.yml
@@ -43,7 +43,7 @@ steps:
     feedsToUse: config
     configPath: NuGet.config
     restoreSolution: src\bug-report\BugReportTool\BugReportTool.sln
-    restoreDirectory: '$(Build.SourcesDirectory)\src\bug-report\BugReportTool\packages'
+    restoreDirectory: '$(Build.SourcesDirectory)\tools\BugReportTool\packages'
 
 - task: VSBuild@1
   displayName: 'Build BugReportTool.sln'
