+++ tools/BugReportTool/BugReportTool/BugReportTool.vcxproj
@@ -30,8 +30,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\..\..\deps\cziplib\src\zip.c" />
-    <ClCompile Include="..\..\..\..\tools\monitor_info_report\report-monitor-info.cpp" />
+    <ClCompile Include="..\..\..\deps\cziplib\src\zip.c" />
+    <ClCompile Include="ReportMonitorInfo.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="ZipTools\zipfolder.cpp" />
   </ItemGroup>
@@ -39,14 +39,14 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\..\common\SettingsAPI\SetttingsAPI.vcxproj">
+    <ProjectReference Include="..\..\..\src\common\SettingsAPI\SetttingsAPI.vcxproj">
       <Project>{6955446d-23f7-4023-9bb3-8657f904af99}</Project>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\..\..\deps\cziplib\src\miniz.h" />
-    <ClInclude Include="..\..\..\..\deps\cziplib\src\zip.h" />
-    <ClInclude Include="..\..\..\..\tools\monitor_info_report\report-monitor-info.h" />
+    <ClInclude Include="..\..\..\deps\cziplib\src\miniz.h" />
+    <ClInclude Include="..\..\..\deps\cziplib\src\zip.h" />
+    <ClInclude Include="ReportMonitorInfo.h" />
     <ClInclude Include="..\..\..\common\utils\json.h" />
     <ClInclude Include="ZipTools\zipfolder.h" />
   </ItemGroup>
