+++ tools/BugReportTool/BugReportTool/ReportMonitorInfo.h
