+++ tools/BugReportTool/BugReportTool/ZipTools/zipfolder.cpp
