+++ tools/BugReportTool/BugReportTool/ZipTools/zipfolder.h
