+++ tools/BugReportTool/BugReportTool/packages.config
